Refactor helpers/session/session.go
