+++ helpers/html/render_html.go
@@ -3,7 +3,7 @@ package html
 import (
 	"github.com/foolin/goview/supports/ginview"
 	"github.com/gin-gonic/gin"
-	session "github.com/gutakk/go-google-scraper/helpers/session"
+	"github.com/gutakk/go-google-scraper/helpers/session"
 )
 
 func RenderWithError(c *gin.Context, status int, view string, title string, err error, data map[string]interface{}) {
